All error fixed

+++ Scissors_backend/src/services/CategoryService.ts
@@ -1,9 +1,13 @@
+import { plainToClass } from "class-transformer";
 import { HttpStatus } from "../constants/HttpStatus";
 import { Messages } from "../constants/Messages";
-import { ICategory, ICategoryDocument } from "../Interfaces/Category/ICategory";
+import { CategoryDto, CreateCategoryDto, UpdateCategoryDto } from "../dto/category.dto";
 import { ICategoryRepository } from "../Interfaces/Category/ICategoryRepository";
 import { ICategoryService } from "../Interfaces/Category/ICategoryService";
 import CustomError from "../Utils/cutsomError";
+import mongoose from "mongoose";
+import { validate } from "class-validator";
+
 
 class CategoryService implements ICategoryService {
   private _repository: ICategoryRepository;
@@ -12,16 +16,22 @@ class CategoryService implements ICategoryService {
     this._repository = repository;
   }
 
-  async getAllCategory(): Promise<ICategoryDocument[]> {
-    const result = await this._repository.getAllCategory();
-    if (!result || result.length === 0) {
+  async getAllCategory(): Promise<CategoryDto[]> {
+   const categories = await this._repository.getAllCategory();
+    if (!categories || categories.length === 0) {
       throw new CustomError(Messages.NO_CATEGORIES_FOUND, HttpStatus.NOT_FOUND);
     }
-    return result;
+    return categories.map((category) =>
+      plainToClass(CategoryDto, {
+        _id: (category._id as mongoose.Types.ObjectId).toString(),
+        name: category.name,
+        description: category.description,
+      })
+    );
   }
 
   async getFilteredCategory(page: number | string, limit: number | string, search: string): Promise<{
-    categories: ICategoryDocument[];
+    categories: CategoryDto[];
     totalItems: number;
     totalPages: number;
     currentPage: number;
@@ -50,44 +60,84 @@ class CategoryService implements ICategoryService {
     const totalPages = Math.ceil(totalItems / limitNumber);
 
     return {
-      categories,
+      categories: categories.map((category) =>
+        plainToClass(CategoryDto, {
+          _id: (category._id as mongoose.Types.ObjectId).toString(),
+          name: category.name,
+          description: category.description,
+        })
+      ),
       totalItems,
       totalPages,
       currentPage: pageNumber,
     };
   }
 
-  async createCategory(categoryData: ICategory): Promise<ICategoryDocument> {
-    if (!categoryData || Object.keys(categoryData).length === 0 || !categoryData.name || !categoryData.description) {
-      throw new CustomError(Messages.INVALID_CATEGORY_DATA, HttpStatus.BAD_REQUEST);
+  async createCategory(categoryData: CreateCategoryDto): Promise<CategoryDto> {
+    const createCategoryDto = plainToClass(CreateCategoryDto,categoryData)
+    const errors = await validate(createCategoryDto)
+    if (errors.length > 0) {
+      throw new CustomError(
+        Messages.INVALID_CATEGORY_DATA + ': ' + errors.map((err) => Object.values(err.constraints || {})).join(', '),
+        HttpStatus.BAD_REQUEST
+      );
     }
-    const result = await this._repository.createCategory(categoryData);
-    return result;
+    const existingCategory = await this._repository.findByName(createCategoryDto.name);
+    if (existingCategory) {
+      throw new CustomError("Category already exists", HttpStatus.CONFLICT);
+    }
+    const result = await this._repository.createCategory({
+      name: createCategoryDto.name,
+      description: createCategoryDto.description,
+    });
+
+    return plainToClass(CategoryDto, {
+      _id: (result._id as mongoose.Types.ObjectId).toString(),
+      name: result.name,
+      description: result.description,
+    });
   }
 
-  async updateCategory(updatedData: {
-    id: string;
-    name: string;
-    description: string;
-  }): Promise<ICategoryDocument | null> {
-    const { id, name, description } = updatedData;
-    if (!id || !name || !description) {
-      throw new CustomError(Messages.INVALID_CATEGORY_DATA, HttpStatus.BAD_REQUEST);
+  async updateCategory(updatedData: UpdateCategoryDto): Promise<CategoryDto | null> {
+
+    const updateCategoryDto = plainToClass(UpdateCategoryDto,updatedData)
+    const errors = await validate(updateCategoryDto)
+   if (errors.length > 0) {
+      throw new CustomError(
+        Messages.INVALID_CATEGORY_DATA + ': ' + errors.map((err) => Object.values(err.constraints || {})).join(', '),
+        HttpStatus.BAD_REQUEST
+      );
     }
-    const category = await this._repository.findByIdCategory(id);
+
+    const category = await this._repository.findByIdCategory(updateCategoryDto.id as string);
     if (!category) {
       throw new CustomError(Messages.CATEGORY_NOT_FOUND, HttpStatus.NOT_FOUND);
     }
-    const result = await this._repository.updateCategory(id, { name, description });
-    return result;
+   const result = await this._repository.updateCategory(updateCategoryDto.id as string, {
+      name: updateCategoryDto.name,
+      description: updateCategoryDto.description,
+    });
+    if (!result) {
+      throw new CustomError(Messages.CATEGORY_NOT_FOUND, HttpStatus.NOT_FOUND);
+    }
+    return plainToClass(CategoryDto, {
+      _id: (result._id as mongoose.Types.ObjectId).toString(),
+      name: result.name,
+      description: result.description,
+    });
   }
 
   async deleteCategory(id: string): Promise<string> {
     if (!id) {
       throw new CustomError(Messages.INVALID_CATEGORY_ID, HttpStatus.BAD_REQUEST);
     }
-    const result = await this._repository.deleteCategory(id);
-    return result;
+   const category = await this._repository.findByIdCategory(id);
+    if (!category) {
+      throw new CustomError(Messages.CATEGORY_NOT_FOUND, HttpStatus.NOT_FOUND);
+    }
+    await this._repository.deleteCategory(id);
+
+    return Messages.CATEGORY_DELETED;
   }
 }
